Move generateOperationSignature method into util method
