+++ packages/core-sdk/src/types/resources/ipAsset.ts
@@ -438,48 +438,6 @@ export type CommonRegistrationTxResponse = RegistrationResponse & {
   txHash: Hex;
 };
 
-export enum SignatureMethodType {
-  REGISTER_IP_AND_MAKE_DERIVATIVE_AND_DEPLOY_ROYALTY_VAULT = "registerIpAndMakeDerivativeAndDeployRoyaltyVault",
-  DISTRIBUTE_ROYALTY_TOKENS = "distributeRoyaltyTokens",
-  REGISTER_DERIVATIVE_IP = "registerDerivativeIp",
-  REGISTER_IP_AND_ATTACH_PIL_TERMS_AND_DEPLOY_ROYALTY_VAULT = "registerIpAndAttachPilTermsAndDeployRoyaltyVault",
-  REGISTER_IP_AND_ATTACH_PIL_TERMS = "registerIpAndAttachPilTerms",
-  REGISTER_IP_AND_MAKE_DERIVATIVE_WITH_LICENSE_TOKENS = "registerIpAndMakeDerivativeWithLicenseTokens",
-  REGISTER_PIL_TERMS_AND_ATTACH = "registerPilTermsAndAttach",
-  REGISTER = "register",
-  BATCH_REGISTER_DERIVATIVE = "batchRegisterDerivative",
-}
-
-export type GeneratePrefixRegisterSignatureRequest = {
-  deadline: bigint;
-  ipIdAddress: Address;
-  methodType: SignatureMethodType;
-  ipRoyaltyVault?: Address;
-  totalAmount?: number;
-  state?: Hex;
-  encodeData?: Hex;
-};
-
-export type MintSpgNftRegistrationRequest = Omit<
-  | MintAndRegisterIpAndMakeDerivativeRequest
-  | MintAndRegisterIpAssetWithPilTermsRequest
-  | MintAndRegisterIpAndAttachPILTermsAndDistributeRoyaltyTokensRequest
-  | MintAndRegisterIpAndMakeDerivativeAndDistributeRoyaltyTokensRequest,
-  "txOptions" | "wipOptions"
->;
-export type RegisterRegistrationRequest = Omit<
-  | RegisterDerivativeAndAttachLicenseTermsAndDistributeRoyaltyTokensRequest
-  | RegisterIpAndAttachPilTermsRequest
-  | RegisterIPAndAttachLicenseTermsAndDistributeRoyaltyTokensRequest
-  | RegisterIpAndMakeDerivativeRequest,
-  "txOptions" | "wipOptions"
->;
-
-export type IpRegistrationWorkflowRequest =
-  | MintSpgNftRegistrationRequest
-  | RegisterRegistrationRequest
-  | DistributeRoyaltyTokens;
-
 export type TransformIpRegistrationWorkflowRequest =
   | RoyaltyTokenDistributionWorkflowsMintAndRegisterIpAndAttachPilTermsAndDistributeRoyaltyTokensRequest
   | LicenseAttachmentWorkflowsMintAndRegisterIpAndAttachPilTermsRequest
@@ -531,3 +489,23 @@ export type ProcessMintAndRegisterRequest = WithWipOptions & {
   >[];
   txOptions?: Omit<WaitForTransactionReceiptParameters, "hash">;
 };
+
+export type MintSpgNftRegistrationRequest = Omit<
+  | MintAndRegisterIpAndMakeDerivativeRequest
+  | MintAndRegisterIpAssetWithPilTermsRequest
+  | MintAndRegisterIpAndAttachPILTermsAndDistributeRoyaltyTokensRequest
+  | MintAndRegisterIpAndMakeDerivativeAndDistributeRoyaltyTokensRequest,
+  "txOptions" | "wipOptions"
+>;
+export type RegisterRegistrationRequest = Omit<
+  | RegisterDerivativeAndAttachLicenseTermsAndDistributeRoyaltyTokensRequest
+  | RegisterIpAndAttachPilTermsRequest
+  | RegisterIPAndAttachLicenseTermsAndDistributeRoyaltyTokensRequest
+  | RegisterIpAndMakeDerivativeRequest,
+  "txOptions" | "wipOptions"
+>;
+
+export type IpRegistrationWorkflowRequest =
+  | MintSpgNftRegistrationRequest
+  | RegisterRegistrationRequest
+  | DistributeRoyaltyTokens;
