+++ packages/core-sdk/src/types/resources/permission.ts
@@ -83,7 +83,7 @@ export type PermissionSignatureRequest = {
   state: Hex;
   /** The deadline for the signature in seconds. */
   deadline: string | number | bigint;
-  wallet: WalletClient;
+  wallet: SimpleWalletClient;
   chainId: string | number | bigint;
   permissions: Omit<SetPermissionsRequest, "txOptions">[];
 };
